用户权限数结构修改

+++ server/model/system/authority.go
@@ -7,10 +7,10 @@ import (
 // Authority 授权
 type Authority struct {
 	model.BaseOrm
-	Id            uint64       `json:"id" form:"id"`                       // 主键ID
+	Id            int          `json:"id" form:"id"`                       // 主键ID
 	Name          string       `json:"name" form:"name"`                   // 角色名
 	Desc          string       `json:"desc" form:"desc"`                   // 简介
-	Pid           uint64       `json:"pid" form:"pid"`                     // 父角色ID
+	Pid           int          `json:"pid" form:"pid"`                     // 父角色ID
 	DefaultRouter string       `json:"defaultRouter" form:"DefaultRouter"` // 默认菜单(默认dashboard)
 	MenuIds       string       `json:"menuIds" form:"menuIds"`             // 权限下的ID
 	ApiIds        string       `json:"apiIds" form:"apiIds"`               // 权限下的ID
